application/requirements.txt
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
mysqlclient==2.0.3
